P4ADEV-2505 storing debtPositionId on SendNotification
