+++ src/main/java/it/gov/pagopa/pu/send/connector/debtpositions/config/DebtPositionApisHolder.java
@@ -1,0 +1,54 @@
+package it.gov.pagopa.pu.send.connector.debtpositions.config;
+
+
+import it.gov.pagopa.pu.debtposition.client.generated.DebtPositionSearchControllerApi;
+import it.gov.pagopa.pu.debtposition.generated.ApiClient;
+import it.gov.pagopa.pu.debtposition.generated.BaseApi;
+import it.gov.pagopa.pu.send.config.RestTemplateConfig;
+import jakarta.annotation.PreDestroy;
+import org.springframework.boot.web.client.RestTemplateBuilder;
+import org.springframework.stereotype.Service;
+import org.springframework.web.client.RestTemplate;
+
+@Service
+public class DebtPositionApisHolder {
+
+  private final DebtPositionSearchControllerApi debtPositionSearchApi;
+
+  private final ThreadLocal<String> bearerTokenHolder = new ThreadLocal<>();
+
+  public DebtPositionApisHolder(DebtPositionApiClientConfig clientConfig, RestTemplateBuilder restTemplateBuilder) {
+    RestTemplate restTemplate = restTemplateBuilder.build();
+    ApiClient apiClient = buildApiClient(restTemplate, clientConfig);
+
+    if (clientConfig.isPrintBodyWhenError()) {
+      restTemplate.setErrorHandler(RestTemplateConfig.bodyPrinterWhenError("DEBT-POSITIONS"));
+    }
+
+    this.debtPositionSearchApi = new DebtPositionSearchControllerApi(apiClient);
+  }
+
+  @PreDestroy
+  public void unload() {
+    bearerTokenHolder.remove();
+  }
+
+  public DebtPositionSearchControllerApi getDebtPositionSearchApi(String accessToken) {
+    return getApi(accessToken, debtPositionSearchApi);
+  }
+
+  private ApiClient buildApiClient(RestTemplate restTemplate, DebtPositionApiClientConfig clientConfig) {
+    ApiClient apiClient = new ApiClient(restTemplate);
+    apiClient.setBasePath(clientConfig.getBaseUrl());
+    apiClient.setBearerToken(bearerTokenHolder::get);
+    apiClient.setMaxAttemptsForRetry(Math.max(1, clientConfig.getMaxAttempts()));
+    apiClient.setWaitTimeMillis(clientConfig.getWaitTimeMillis());
+    return apiClient;
+  }
+
+  private <T extends BaseApi> T getApi(String accessToken, T api) {
+    bearerTokenHolder.set(accessToken);
+    return api;
+  }
+}
+
