+++ src/main/java/it/gov/pagopa/pu/send/connector/organization/client/OrganizationApiClient.java
@@ -1,0 +1,29 @@
+package it.gov.pagopa.pu.send.connector.organization.client;
+
+import it.gov.pagopa.pu.organization.dto.generated.OrganizationApiKeyType;
+import it.gov.pagopa.pu.send.connector.organization.config.OrganizationApisHolder;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+import org.springframework.web.client.HttpClientErrorException;
+
+@Slf4j
+@Service
+public class OrganizationApiClient {
+
+  private final OrganizationApisHolder organizationApisHolder;
+
+  public OrganizationApiClient(OrganizationApisHolder organizationApisHolder) {
+    this.organizationApisHolder = organizationApisHolder;
+  }
+
+  public String getOrganizationApiKey(Long organizationId, String accessToken) {
+    try{
+      return organizationApisHolder.getOrganizationApi(accessToken)
+        .getOrganizationApiKey(organizationId, OrganizationApiKeyType.SEND);
+    } catch (HttpClientErrorException.NotFound e){
+      log.info("Cannot find organization having organizationId {}", organizationId);
+      return null;
+    }
+  }
+
+}
