+++ src/main/java/it/gov/pagopa/pu/send/connector/organization/config/OrganizationApiClientConfig.java
@@ -1,0 +1,14 @@
+package it.gov.pagopa.pu.send.connector.organization.config;
+
+import it.gov.pagopa.pu.send.config.ApiClientConfig;
+import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@ConfigurationProperties(prefix = "rest.organization")
+@SuperBuilder
+@NoArgsConstructor
+public class OrganizationApiClientConfig extends ApiClientConfig {
+}
