+++ src/main/java/it/gov/pagopa/pu/send/connector/organization/service/OrganizationService.java
@@ -1,0 +1,6 @@
+package it.gov.pagopa.pu.send.connector.organization.service;
+
+public interface OrganizationService {
+
+  String getOrganizationApiKey(Long organizationId, String accessToken);
+}
