+++ src/main/java/it/gov/pagopa/pu/send/connector/organization/service/OrganizationServiceImpl.java
@@ -1,0 +1,19 @@
+package it.gov.pagopa.pu.send.connector.organization.service;
+
+import it.gov.pagopa.pu.send.connector.organization.client.OrganizationApiClient;
+import org.springframework.stereotype.Service;
+
+@Service
+public class OrganizationServiceImpl implements OrganizationService {
+
+  private final OrganizationApiClient organizationApiClient;
+
+  public OrganizationServiceImpl(OrganizationApiClient organizationApiClient) {
+    this.organizationApiClient = organizationApiClient;
+  }
+
+  @Override
+  public String getOrganizationApiKey(Long organizationId, String accessToken) {
+    return organizationApiClient.getOrganizationApiKey(organizationId, accessToken);
+  }
+}
