--- src/main/java/it/gov/pagopa/pu/send/connector/pagopa/organization/config/OrganizationApisHolder.java
@@ -1,53 +1,0 @@
-package it.gov.pagopa.pu.send.connector.pagopa.organization.config;
-
-
-import it.gov.pagopa.pu.organization.client.generated.OrganizationApi;
-import it.gov.pagopa.pu.organization.generated.ApiClient;
-import it.gov.pagopa.pu.organization.generated.BaseApi;
-import it.gov.pagopa.pu.send.config.RestTemplateConfig;
-import jakarta.annotation.PreDestroy;
-import org.springframework.boot.web.client.RestTemplateBuilder;
-import org.springframework.stereotype.Service;
-import org.springframework.web.client.RestTemplate;
-
-@Service
-public class OrganizationApisHolder {
-
-  private final OrganizationApi organizationApi;
-  private final ThreadLocal<String> bearerTokenHolder = new ThreadLocal<>();
-
-  public OrganizationApisHolder(OrganizationApiClientConfig clientConfig, RestTemplateBuilder restTemplateBuilder) {
-    RestTemplate restTemplate = restTemplateBuilder.build();
-    ApiClient apiClient = buildApiClient(restTemplate, clientConfig);
-
-    if (clientConfig.isPrintBodyWhenError()) {
-      restTemplate.setErrorHandler(RestTemplateConfig.bodyPrinterWhenError("ORGANIZATION"));
-    }
-
-    this.organizationApi = new OrganizationApi(apiClient);
-  }
-
-  @PreDestroy
-  public void unload() {
-    bearerTokenHolder.remove();
-  }
-
-  public OrganizationApi getOrganizationApi(String accessToken) {
-    return getApi(accessToken, organizationApi);
-  }
-
-  private ApiClient buildApiClient(RestTemplate restTemplate, OrganizationApiClientConfig clientConfig) {
-    ApiClient apiClient = new ApiClient(restTemplate);
-    apiClient.setBasePath(clientConfig.getBaseUrl());
-    apiClient.setBearerToken(bearerTokenHolder::get);
-    apiClient.setMaxAttemptsForRetry(Math.max(1, clientConfig.getMaxAttempts()));
-    apiClient.setWaitTimeMillis(clientConfig.getWaitTimeMillis());
-    return apiClient;
-  }
-
-  private <T extends BaseApi> T getApi(String accessToken, T api) {
-    bearerTokenHolder.set(accessToken);
-    return api;
-  }
-}
-
