--- src/main/java/it/gov/pagopa/pu/send/connector/pagopa/organization/service/OrganizationService.java
@@ -1,6 +1,0 @@
-package it.gov.pagopa.pu.send.connector.pagopa.organization.service;
-
-public interface OrganizationService {
-
-  String getOrganizationApiKey(Long organizationId, String accessToken);
-}
