--- src/main/java/it/gov/pagopa/pu/send/connector/pagopa/workflow/client/WorkflowApiClient.java
@@ -1,20 +1,0 @@
-package it.gov.pagopa.pu.send.connector.pagopa.workflow.client;
-
-import it.gov.pagopa.pu.send.connector.pagopa.workflow.config.WorkflowApisHolder;
-import it.gov.pagopa.pu.workflowhub.dto.generated.WorkflowCreatedDTO;
-import org.springframework.stereotype.Service;
-
-@Service
-public class WorkflowApiClient {
-
-  private final WorkflowApisHolder workflowApisHolder;
-
-  public WorkflowApiClient(WorkflowApisHolder workflowApisHolder) {
-    this.workflowApisHolder = workflowApisHolder;
-  }
-
-  public WorkflowCreatedDTO sendNotificationProcess(String sendNotificationId, String accessToken) {
-    return workflowApisHolder.getSendNotificationApi(accessToken)
-      .sendNotificationProcess(sendNotificationId);
-  }
-}
