--- src/main/java/it/gov/pagopa/pu/send/connector/pagopa/workflow/config/WorkflowApiClientConfig.java
@@ -1,14 +1,0 @@
-package it.gov.pagopa.pu.send.connector.pagopa.workflow.config;
-
-import it.gov.pagopa.pu.send.config.ApiClientConfig;
-import lombok.NoArgsConstructor;
-import lombok.experimental.SuperBuilder;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@ConfigurationProperties(prefix = "rest.pagopa.workflow-hub")
-@SuperBuilder
-@NoArgsConstructor
-public class WorkflowApiClientConfig extends ApiClientConfig {
-}
