+++ src/main/java/it/gov/pagopa/pu/send/connector/workflow/config/WorkflowApiClientConfig.java
@@ -1,0 +1,14 @@
+package it.gov.pagopa.pu.send.connector.workflow.config;
+
+import it.gov.pagopa.pu.send.config.ApiClientConfig;
+import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@ConfigurationProperties(prefix = "rest.workflow-hub")
+@SuperBuilder
+@NoArgsConstructor
+public class WorkflowApiClientConfig extends ApiClientConfig {
+}
