+++ src/main/java/it/gov/pagopa/pu/send/connector/workflow/service/WorkflowService.java
@@ -1,0 +1,8 @@
+package it.gov.pagopa.pu.send.connector.workflow.service;
+
+import it.gov.pagopa.pu.workflowhub.dto.generated.WorkflowCreatedDTO;
+
+public interface WorkflowService {
+
+  WorkflowCreatedDTO sendNotificationProcess(String sendNotificationId, String accessToken);
+}
