+++ src/main/java/it/gov/pagopa/pu/send/dto/PuPayment.java
@@ -1,0 +1,16 @@
+package it.gov.pagopa.pu.send.dto;
+
+import it.gov.pagopa.pu.send.dto.generated.Payment;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@SuperBuilder
+@Data
+public class PuPayment {
+  private Long debtPositionId;
+  private Payment payment;
+}
