+++ src/main/java/it/gov/pagopa/pu/send/exception/UnknownDebtPositionException.java
@@ -1,0 +1,7 @@
+package it.gov.pagopa.pu.send.exception;
+
+public class UnknownDebtPositionException extends RuntimeException {
+  public UnknownDebtPositionException(String message){
+    super(message);
+  }
+}
