+++ src/test/java/it/gov/pagopa/pu/send/connector/debtpositions/config/DebtPositionApisHolderTest.java
@@ -1,0 +1,47 @@
+package it.gov.pagopa.pu.send.connector.debtpositions.config;
+
+import it.gov.pagopa.pu.send.connector.BaseApiHolderTest;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mock;
+import org.mockito.Mockito;
+import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.boot.web.client.RestTemplateBuilder;
+import org.springframework.core.ParameterizedTypeReference;
+import org.springframework.web.util.DefaultUriBuilderFactory;
+
+@ExtendWith(MockitoExtension.class)
+class DebtPositionApisHolderTest extends BaseApiHolderTest {
+  @Mock
+  private RestTemplateBuilder restTemplateBuilderMock;
+
+  private DebtPositionApisHolder apisHolder;
+
+  @BeforeEach
+  void setUp() {
+    Mockito.when(restTemplateBuilderMock.build()).thenReturn(restTemplateMock);
+    Mockito.when(restTemplateMock.getUriTemplateHandler()).thenReturn(new DefaultUriBuilderFactory());
+    DebtPositionApiClientConfig apiClient = new DebtPositionApiClientConfig();
+    apiClient.setBaseUrl("http://example.com");
+    apisHolder = new DebtPositionApisHolder(apiClient, restTemplateBuilderMock);
+  }
+
+  @AfterEach
+  void verifyNoMoreInteractions() {
+    Mockito.verifyNoMoreInteractions(
+      restTemplateBuilderMock,
+      restTemplateMock
+    );
+  }
+
+  @Test
+  void whenGetOrganizationApiThenAuthenticationShouldBeSetInThreadSafeMode() throws InterruptedException {
+    assertAuthenticationShouldBeSetInThreadSafeMode(
+      token -> apisHolder.getDebtPositionSearchApi(token)
+        .crudDebtPositionsFindByOrganizationIdAndInstallmentNav(1L, "NAV"),
+      new ParameterizedTypeReference<>() {},
+      apisHolder::unload);
+  }
+}
