+++ src/test/java/it/gov/pagopa/pu/send/connector/debtpositions/service/DebtPositionServiceTest.java
@@ -1,0 +1,50 @@
+package it.gov.pagopa.pu.send.connector.debtpositions.service;
+
+import it.gov.pagopa.pu.debtposition.dto.generated.DebtPosition;
+import it.gov.pagopa.pu.send.connector.debtpositions.client.DebtPositionSearchApiClient;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mock;
+import org.mockito.Mockito;
+import org.mockito.junit.jupiter.MockitoExtension;
+
+@ExtendWith(MockitoExtension.class)
+class DebtPositionServiceTest {
+
+  @Mock
+  private DebtPositionSearchApiClient debtPositionSearchApiClientMock;
+
+  private DebtPositionService service;
+
+  @BeforeEach
+  void init(){
+    service = new DebtPositionServiceImpl(debtPositionSearchApiClientMock);
+  }
+
+  @AfterEach
+  void verifyNoMoreInteractions(){
+    Mockito.verifyNoMoreInteractions(debtPositionSearchApiClientMock);
+  }
+
+  @Test
+  void whenGetOrganizationApiKeyThenInvokeClient(){
+    // Given
+    Long organizationId = 1L;
+    String nav = "NAV";
+    String accessToken = "accessToken";
+
+    DebtPosition expectedResult = new DebtPosition();
+
+    Mockito.when(debtPositionSearchApiClientMock.findDebtPositionByInstallment(organizationId, nav, accessToken))
+      .thenReturn(expectedResult);
+
+    // When
+    DebtPosition result = service.findDebtPositionByInstallment(organizationId, nav, accessToken);
+
+    // Then
+    Assertions.assertSame(expectedResult, result);
+  }
+}
