+++ src/test/java/it/gov/pagopa/pu/send/connector/organization/config/OrganizationApisHolderTest.java
@@ -1,0 +1,48 @@
+package it.gov.pagopa.pu.send.connector.organization.config;
+
+import it.gov.pagopa.pu.organization.dto.generated.OrganizationApiKeyType;
+import it.gov.pagopa.pu.send.connector.BaseApiHolderTest;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mock;
+import org.mockito.Mockito;
+import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.boot.web.client.RestTemplateBuilder;
+import org.springframework.core.ParameterizedTypeReference;
+import org.springframework.web.util.DefaultUriBuilderFactory;
+
+@ExtendWith(MockitoExtension.class)
+class OrganizationApisHolderTest extends BaseApiHolderTest {
+  @Mock
+  private RestTemplateBuilder restTemplateBuilderMock;
+
+  private OrganizationApisHolder apisHolder;
+
+  @BeforeEach
+  void setUp() {
+    Mockito.when(restTemplateBuilderMock.build()).thenReturn(restTemplateMock);
+    Mockito.when(restTemplateMock.getUriTemplateHandler()).thenReturn(new DefaultUriBuilderFactory());
+    OrganizationApiClientConfig apiClient = new OrganizationApiClientConfig();
+    apiClient.setBaseUrl("http://example.com");
+    apisHolder = new OrganizationApisHolder(apiClient, restTemplateBuilderMock);
+  }
+
+  @AfterEach
+  void verifyNoMoreInteractions() {
+    Mockito.verifyNoMoreInteractions(
+      restTemplateBuilderMock,
+      restTemplateMock
+    );
+  }
+
+  @Test
+  void whenGetOrganizationApiThenAuthenticationShouldBeSetInThreadSafeMode() throws InterruptedException {
+    assertAuthenticationShouldBeSetInThreadSafeMode(
+      token -> apisHolder.getOrganizationApi(token)
+        .getOrganizationApiKey(1L, OrganizationApiKeyType.SEND),
+      new ParameterizedTypeReference<>() {},
+      apisHolder::unload);
+  }
+}
