--- src/test/java/it/gov/pagopa/pu/send/connector/pagopa/workflow/config/WorkflowApisHolderTest.java
@@ -1,49 +1,0 @@
-package it.gov.pagopa.pu.send.connector.pagopa.workflow.config;
-
-import it.gov.pagopa.pu.send.connector.BaseApiHolderTest;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.Mock;
-import org.mockito.Mockito;
-import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.boot.web.client.RestTemplateBuilder;
-import org.springframework.core.ParameterizedTypeReference;
-import org.springframework.web.util.DefaultUriBuilderFactory;
-
-@ExtendWith(MockitoExtension.class)
-class WorkflowApisHolderTest extends BaseApiHolderTest {
-    @Mock
-    private RestTemplateBuilder restTemplateBuilderMock;
-
-    private WorkflowApisHolder workflowApisHolder;
-
-    @BeforeEach
-    void setUp() {
-        Mockito.when(restTemplateBuilderMock.build()).thenReturn(restTemplateMock);
-        Mockito.when(restTemplateMock.getUriTemplateHandler()).thenReturn(new DefaultUriBuilderFactory());
-        WorkflowApiClientConfig clientConfig = WorkflowApiClientConfig.builder()
-          .baseUrl("http://example.com")
-          .build();
-        workflowApisHolder = new WorkflowApisHolder(clientConfig, restTemplateBuilderMock);
-    }
-
-    @AfterEach
-    void verifyNoMoreInteractions() {
-        Mockito.verifyNoMoreInteractions(
-                restTemplateBuilderMock,
-                restTemplateMock
-        );
-    }
-
-    @Test
-    void whenSendNotificationApiThenAuthenticationShouldBeSetInThreadSafeMode() throws InterruptedException {
-        assertAuthenticationShouldBeSetInThreadSafeMode(
-                accessToken -> workflowApisHolder.getSendNotificationApi(accessToken)
-                  .sendNotificationProcess("sendNotificationId"),
-                new ParameterizedTypeReference<>() {},
-                workflowApisHolder::unload);
-    }
-
-}
