--- src/test/java/it/gov/pagopa/pu/send/connector/pagopa/workflow/service/WorkflowServiceTest.java
@@ -1,53 +1,0 @@
-package it.gov.pagopa.pu.send.connector.pagopa.workflow.service;
-
-import it.gov.pagopa.pu.send.connector.pagopa.workflow.client.WorkflowApiClient;
-import it.gov.pagopa.pu.workflowhub.dto.generated.WorkflowCreatedDTO;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.Mock;
-import org.mockito.Mockito;
-import org.mockito.junit.jupiter.MockitoExtension;
-
-@ExtendWith(MockitoExtension.class)
-class WorkflowServiceTest {
-
-  @Mock
-  private WorkflowApiClient workflowApiClientMock;
-
-  private WorkflowService workflowService;
-
-  @BeforeEach
-  void init() {
-    workflowService = new WorkflowServiceImpl(
-      workflowApiClientMock
-    );
-  }
-
-  @AfterEach
-  void verifyNoMoreInteractions() {
-    Mockito.verifyNoMoreInteractions(
-      workflowApiClientMock
-    );
-  }
-
-  @Test
-  void whenSendNotificationProcessThenOk() {
-    // Given
-    String accessToken = "ACCESSTOKEN";
-    String sendNotificationId = "sendNotificationId";
-    WorkflowCreatedDTO expectedResult = new WorkflowCreatedDTO("1");
-
-    Mockito.when(workflowApiClientMock.sendNotificationProcess(Mockito.same(sendNotificationId), Mockito.same(accessToken)))
-      .thenReturn(expectedResult);
-
-    // When
-    WorkflowCreatedDTO result = workflowService.sendNotificationProcess(sendNotificationId, accessToken);
-
-    // Then
-    Assertions.assertSame(expectedResult, result);
-  }
-
-}
