+++ src/test/java/it/gov/pagopa/pu/send/mapper/CreateNotificationRequest2SendNotificationMapperTest.java
@@ -1,35 +1,87 @@
 package it.gov.pagopa.pu.send.mapper;
 
-import it.gov.pagopa.pu.send.dto.generated.Attachment;
-import it.gov.pagopa.pu.send.dto.generated.CreateNotificationRequest;
+import it.gov.pagopa.pu.debtposition.dto.generated.DebtPosition;
+import it.gov.pagopa.pu.send.connector.debtpositions.service.DebtPositionService;
+import it.gov.pagopa.pu.send.dto.generated.*;
 import it.gov.pagopa.pu.send.dto.generated.CreateNotificationRequest.NotificationFeePolicyEnum;
 import it.gov.pagopa.pu.send.dto.generated.CreateNotificationRequest.PagoPaIntModeEnum;
 import it.gov.pagopa.pu.send.dto.generated.CreateNotificationRequest.PhysicalCommunicationTypeEnum;
-import it.gov.pagopa.pu.send.dto.generated.Document;
-import it.gov.pagopa.pu.send.dto.generated.PagoPa;
-import it.gov.pagopa.pu.send.dto.generated.Payment;
-import it.gov.pagopa.pu.send.dto.generated.Recipient;
 import it.gov.pagopa.pu.send.dto.generated.Recipient.RecipientTypeEnum;
 import it.gov.pagopa.pu.send.enums.FileStatus;
 import it.gov.pagopa.pu.send.enums.NotificationStatus;
+import it.gov.pagopa.pu.send.exception.UnknownDebtPositionException;
 import it.gov.pagopa.pu.send.model.SendNotification;
 import it.gov.pagopa.pu.send.util.TestUtils;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mock;
+import org.mockito.Mockito;
+import org.mockito.junit.jupiter.MockitoExtension;
+
 import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Collections;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.junit.jupiter.MockitoExtension;
 
 @ExtendWith(MockitoExtension.class)
 class CreateNotificationRequest2SendNotificationMapperTest {
 
-  private final CreateNotificationRequest2SendNotificationMapper mapper = new CreateNotificationRequest2SendNotificationMapper();
+  @Mock
+  private DebtPositionService debtPositionServiceMock;
+
+  private CreateNotificationRequest2SendNotificationMapper mapper;
+
+  @BeforeEach
+  void init(){
+    mapper = new CreateNotificationRequest2SendNotificationMapper(debtPositionServiceMock);
+  }
+
+  @AfterEach
+  void verifyNoMoreInteractions(){
+    Mockito.verifyNoMoreInteractions(debtPositionServiceMock);
+  }
 
   @Test
   void givenCreateNotificationRequestWhenMapThenOk(){
     // Given
+    CreateNotificationRequest request = buildRequest();
+    String nav = request.getRecipient().getPayments().getFirst().getPagoPa().getNoticeCode();
+
+    Long organizationId = 1L;
+    String accessToken = "ACCESSTOKEN";
+
+    DebtPosition debtPosition = new DebtPosition();
+    debtPosition.setDebtPositionId(3L);
+    Mockito.when(debtPositionServiceMock.findDebtPositionByInstallment(organizationId, nav, accessToken))
+      .thenReturn(debtPosition);
+
+    // When
+    SendNotification result = mapper.map(request, organizationId, accessToken);
+
+    // Then
+    TestUtils.checkNotNullFields(result, "sendNotificationId","organizationId","notificationRequestId","iun");
+
+    Assertions.assertNotNull(result);
+    Assertions.assertEquals("PF", result.getSubjectType());
+    Assertions.assertEquals("RSSMRA80L05F593A", result.getFiscalCode());
+    Assertions.assertEquals("ROSSI MARIO", result.getDenomination());
+    checkPayments(debtPosition, result);
+    checkDocuments(result);
+    Assertions.assertEquals(NotificationFeePolicyEnum.DELIVERY_MODE, NotificationFeePolicyEnum.valueOf(result.getNotificationFeePolicy()));
+    Assertions.assertEquals(PhysicalCommunicationTypeEnum.AR_REGISTERED_LETTER, PhysicalCommunicationTypeEnum.valueOf(result.getPhysicalCommunicationType()));
+    Assertions.assertEquals("SENDERDENOMINATION", result.getSenderDenomination());
+    Assertions.assertEquals("TAXID", result.getSenderTaxId());
+    Assertions.assertEquals(99999999, result.getAmount());
+    Assertions.assertEquals("TAXONOMYCODE", result.getTaxonomyCode());
+    Assertions.assertEquals(100, result.getPaFee());
+    Assertions.assertEquals(22, result.getVat());
+    Assertions.assertEquals(LocalDate.now().toString(), result.getPaymentExpirationDate());
+    Assertions.assertEquals(PagoPaIntModeEnum.NONE, PagoPaIntModeEnum.valueOf(result.getPagoPaIntMode()));
+  }
+
+  private static CreateNotificationRequest buildRequest() {
     Recipient recipient = new Recipient();
     recipient.setRecipientType(RecipientTypeEnum.PF);
     recipient.setTaxId("RSSMRA80L05F593A");
@@ -67,22 +119,17 @@ class CreateNotificationRequest2SendNotificationMapperTest {
     request.setVat(22);
     request.setPaymentExpirationDate(LocalDate.now());
     request.setPagoPaIntMode(PagoPaIntModeEnum.NONE);
+    return request;
+  }
 
-    Long organizationId = 1L;
+  private void checkPayments(DebtPosition debtPosition, SendNotification result) {
+    Assertions.assertSame(debtPosition.getDebtPositionId(), result.getPayments().getFirst().getDebtPositionId());
+    Assertions.assertEquals("CREDITORTAXID", result.getPayments().getFirst().getPayment().getPagoPa().getCreditorTaxId());
+    Assertions.assertEquals("NOTICECODE", result.getPayments().getFirst().getPayment().getPagoPa().getNoticeCode());
+    Assertions.assertEquals(true, result.getPayments().getFirst().getPayment().getPagoPa().getApplyCost());
+  }
 
-    // When
-    SendNotification result = mapper.map(request, organizationId);
-
-    // Then
-    TestUtils.checkNotNullFields(result, "sendNotificationId","organizationId","notificationRequestId","iun");
-
-    Assertions.assertNotNull(result);
-    Assertions.assertEquals("PF", result.getSubjectType());
-    Assertions.assertEquals("RSSMRA80L05F593A", result.getFiscalCode());
-    Assertions.assertEquals("ROSSI MARIO", result.getDenomination());
-    Assertions.assertEquals("CREDITORTAXID", result.getPayments().getFirst().getPagoPa().getCreditorTaxId());
-    Assertions.assertEquals("NOTICECODE", result.getPayments().getFirst().getPagoPa().getNoticeCode());
-    Assertions.assertEquals(true, result.getPayments().getFirst().getPagoPa().getApplyCost());
+  private void checkDocuments(SendNotification result) {
     Assertions.assertEquals("attachment.pdf", result.getDocuments().getFirst().getFileName());
     Assertions.assertEquals("application/pdf", result.getDocuments().getFirst().getContentType());
     Assertions.assertEquals("sha256", result.getDocuments().getFirst().getDigest());
@@ -91,15 +138,19 @@ class CreateNotificationRequest2SendNotificationMapperTest {
     Assertions.assertEquals("sha256", result.getDocuments().getLast().getDigest());
     Assertions.assertEquals(NotificationStatus.WAITING_FILE, result.getStatus());
     Assertions.assertEquals(FileStatus.WAITING, result.getDocuments().getFirst().getStatus());
-    Assertions.assertEquals(NotificationFeePolicyEnum.DELIVERY_MODE, NotificationFeePolicyEnum.valueOf(result.getNotificationFeePolicy()));
-    Assertions.assertEquals(PhysicalCommunicationTypeEnum.AR_REGISTERED_LETTER, PhysicalCommunicationTypeEnum.valueOf(result.getPhysicalCommunicationType()));
-    Assertions.assertEquals("SENDERDENOMINATION", result.getSenderDenomination());
-    Assertions.assertEquals("TAXID", result.getSenderTaxId());
-    Assertions.assertEquals(99999999, result.getAmount());
-    Assertions.assertEquals("TAXONOMYCODE", result.getTaxonomyCode());
-    Assertions.assertEquals(100, result.getPaFee());
-    Assertions.assertEquals(22, result.getVat());
-    Assertions.assertEquals(LocalDate.now().toString(), result.getPaymentExpirationDate());
-    Assertions.assertEquals(PagoPaIntModeEnum.NONE, PagoPaIntModeEnum.valueOf(result.getPagoPaIntMode()));
+  }
+
+  @Test
+  void givenNoDebtPositionWhenThenThrow(){
+    Long organizationId = 1L;
+    String accessToken = "ACCESSTOKEN";
+    CreateNotificationRequest request = buildRequest();
+    String nav = request.getRecipient().getPayments().getFirst().getPagoPa().getNoticeCode();
+
+    Mockito.when(debtPositionServiceMock.findDebtPositionByInstallment(organizationId, nav, accessToken))
+      .thenReturn(null);
+
+    // When, Then
+    Assertions.assertThrows(UnknownDebtPositionException.class, () -> mapper.map(request, organizationId, accessToken));
   }
 }
